Added minimal type panels
