+++ .eslintignore
@@ -39,7 +39,10 @@ src/Tooltip.vue
 src/Typeahead.vue
 
 # MarkBind/vue-strap
-src/PanelSwitch.vue
+src/panels/PanelBase.js
+src/panels/PanelSwitch.vue
+src/panels/NestedPanel.vue
+src/panels/MinimalPanel.vue
 src/Pic.vue
 src/Question.vue
 src/Retriever.vue
